`refactor`: share id to each chip and fix reviews

+++ features/communication/shareexperience/src/main/java/com/tokopedia/shareexperience/data/dto/affiliate/generatelink/ShareExGenerateAffiliateLinkDataResponseDto.kt
@@ -19,7 +19,7 @@ data class ShareExGenerateAffiliateLinkDataResponseDto(
     val identifierType: Int = 0,
 
     @SerializedName("LinkID")
-    val linkID: Int = 0,
+    val linkID: Long = 0,
 
     @SerializedName("URL")
     val url: ShareExGenerateAffiliateLinkUrlResponseDto = ShareExGenerateAffiliateLinkUrlResponseDto()
